Almost finished the implementation of a non-deletable map.
Discovered a problem: an Obj might be storing a reference to the
space that it came from. However, the type of the space may be
different from the type of the object, which is now not allowed
and gives a problem in combination with the map.

Also, need to rethink the approach for avoiding type signatures.